+++ Code/Plugin.cs
@@ -13,7 +13,7 @@ namespace LordsStarstormEdits
         public const string PluginGUID = PluginAuthor + "." + PluginName;
         public const string PluginAuthor = "LordVGames";
         public const string PluginName = "LordsStarstormEdits";
-        public const string PluginVersion = "1.0.1";
+        public const string PluginVersion = "1.1.0";
         public void Awake()
         {
             PluginInfo = Info;