+++ Thunderstore/Package Files/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+- Added option to configure toxic elites off
+- Gave Scavenger's Fortune and Seismic Oscillator an actual description
+- Prevents empyrean elites from spawning in enemies returns' judgement stage
+- Restored most shard drops
+- Made Ultra elite healing a separate config option
+
 ## 1.0.1
 
 - Fixed a few edits preventing the game from loading if SS2's beta is turned off